Move logos into their own files
