+++ packages/web-components/src/LogoWithoutText.tsx
@@ -1,0 +1,38 @@
+import React, { ButtonHTMLAttributes, useMemo } from 'react'
+interface LogoWithoutTextProps extends ButtonHTMLAttributes<HTMLButtonElement> {
+    themeColor: 'light' | 'dark'
+}
+
+export const LogoWithoutText: React.FC<LogoWithoutTextProps> = ({ themeColor }: LogoWithoutTextProps) => {
+    const fill = useMemo(() => (themeColor === 'light' ? '#1d1d1b' : '#fff'), [themeColor])
+    return (
+        <svg
+            className="logo"
+            id="logo-without-text"
+            xmlns="http://www.w3.org/2000/svg"
+            viewBox="0 0 260 348"
+            height="35px"
+            style={{ fill }}
+        >
+            <title>Prosopo Logo Without Text</title>
+            <path
+                id="Vector"
+                d="M95.7053 40.2707C127.005 40.2707 157.022 52.6841 179.154 74.78C201.286 96.8759 213.719 126.844 213.719 158.093H254.056C254.056 70.7808 183.16 -4.57764e-05 95.7053 -4.57764e-05V40.2707Z"
+            />
+            <path
+                id="Vector_2"
+                d="M42.8365 93.0614C58.3333 93.0614 73.6784 96.1087 87.9955 102.029C102.313 107.95 115.322 116.628 126.279 127.568C137.237 138.508 145.93 151.496 151.86 165.79C157.79 180.084 160.843 195.404 160.843 210.875H201.179C201.179 123.564 130.291 52.7906 42.8365 52.7906V93.0614Z"
+            />
+            <path
+                id="Vector_3"
+                d="M158.367 305.005C127.07 305.003 97.056 292.59 74.926 270.496C52.796 248.402 40.3626 218.437 40.3604 187.191H0.0239563C0.0239563 274.503 70.9123 345.276 158.367 345.276V305.005Z"
+            />
+            <path
+                id="Vector_4"
+                d="M211.219 252.239C195.722 252.239 180.376 249.191 166.059 243.27C151.741 237.349 138.732 228.67 127.774 217.729C116.816 206.788 108.123 193.799 102.194 179.505C96.2637 165.21 93.2121 149.889 93.2132 134.417H52.8687C52.8687 221.729 123.765 292.509 211.219 292.509V252.239Z"
+            />
+        </svg>
+    )
+}
+
+export default LogoWithoutText
